// src/upload/upload.enums.ts
export enum FileType {
  PROFILE_IMAGE = 'PROFILE_IMAGE',
  BANNER_IMAGE = 'BANNER_IMAGE',
  POST_IMAGE = 'POST_IMAGE',
  POST_VIDEO = 'POST_VIDEO',
  REVIEW_IMAGE = 'REVIEW_IMAGE',
  REVIEW_VIDEO = 'REVIEW_VIDEO',
  DOCUMENT = 'DOCUMENT',
  BRAND_VERIFICATION = 'BRAND_VERIFICATION',
  MESSAGE_IMAGE = 'MESSAGE_IMAGE',
  MESSAGE_DOCUMENT = 'MESSAGE_DOCUMENT',
}